requirements.txt
antlr4-python3-runtime==4.7.1
awscli==1.14.21
botocore==1.8.25
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.7
cycler==0.10.0
gmpy==1.17
idna==2.6
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
primesieve==1.3.0
pyasn1==0.4.2
pycrypto==2.6.1
pyparsing==2.2.0
pyquil==1.8.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
typing==3.6.4
urllib3==1.22
